requirements.txt
dash
dash-bootstrap-components
spacy
numpy
pandas
umap-learn
tqdm
gensim